add pairwise distance docs and unittest

+++ oneflow/core/functional/impl/nn_functor.cpp
@@ -2633,11 +2633,12 @@ class OneHotFunctor {
 };
 
 class PairwiseDistanceFunctor {
-  public:
-  Maybe<Tensor> operator()(const std::shared_ptr<Tensor>& x, const std::shared_ptr<Tensor>& y, const double& p, const double& eps, bool keepdim) const {
+ public:
+  Maybe<Tensor> operator()(const std::shared_ptr<Tensor>& x, const std::shared_ptr<Tensor>& y,
+                           const float& p, const double& eps, bool keepdim) const {
     int64_t xdim = x->ndim();
     int64_t ydim = y->ndim();
-    int64_t output_dim = xdim > ydim ? xdim: ydim;
+    int64_t output_dim = xdim > ydim ? xdim : ydim;
     auto sub = JUST(ScalarAdd(JUST(Sub(x, y, 1, false)), eps, 1, false));
     return ScalarNorm(sub, p, output_dim - 1, keepdim, NullOpt);
   }
